Reorganize tests into test/ directory with individual test files for each module

+++ test.sh
@@ -3,19 +3,25 @@
 abrt() { exit 1 ; }
 trap abrt ABRT
 
-for file in cz_*.c ; do
-    f="${file%.*}"
-    echo "$file..."
-    cc -std=c2x -I . \
-        -W -Werror -Wall -Wstrict-prototypes \
-        -g -funroll-loops -O3 \
-        -lc -lm \
-    $file -static -o $f && ./$f && {
-        rm -f "./$f"
-    } || {
-        rm -f "./$f"
-        exit 1
-    }
+echo "Running tests from test/ directory..."
+for file in test/*_test.c ; do
+    if [ -f "$file" ]; then
+        basename="${file##*/}"
+        f="test/${basename%.*}"
+        echo "  $basename..."
+        cc -std=c2x -I . \
+            -W -Werror -Wall -Wstrict-prototypes \
+            -g -funroll-loops -O3 \
+            -lc -lm \
+        "$file" -static -o "$f" && "./$f" && {
+            rm -f "./$f"
+        } || {
+            rm -f "./$f"
+            exit 1
+        }
+    fi
 done
+
+echo "All tests passed!"
 
 # EOF
